backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-dotenv==1.0.1
python-multipart==0.0.9
email-validator==2.1.1
httpx==0.27.0
requests>=2.32.0
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Optional: for real email sending (uncommented for Gmail SMTP)
fastapi-mail==1.4.1
aiosmtplib>=2.0.0
